# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PeakNovel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books/{bookId}/chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters of a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChapterSummary"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{id}:
    get:
      operationId: getChapter
      tags: [chapters]
      summary: Read a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterRead"
        "403":
          description: Premium required
        "404":
          description: Not found
    patch:
      operationId: updateChapter
      tags: [chapters]
      summary: Update a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── COMMENTS ───────────────────────────────────────────
  /chapters/{chapterId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: Get comments for a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /comments/{id}/like:
    post:
      operationId: likeComment
      tags: [comments]
      summary: Like a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked

  # ── USERS ──────────────────────────────────────────────
...
  /premium/checkout:
    post:
      operationId: createCheckout
      tags: [premium]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"

  # ── ADMIN ──────────────────────────────────────────────
...
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ChapterSummary:
      type: object
      required: [id, bookId, title, order, isPremium, createdAt]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        title:
          type: string
        order:
          type: integer
        isPremium:
          type: boolean
        wordCount:
          type: integer
        createdAt:
          type: string
    Chapter:
      type: object
      required: [id, bookId, title, order, isPremium, createdAt]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        order:
          type: integer
        isPremium:
          type: boolean
        wordCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ChapterRead:
      type: object
      required: [id, bookId, bookTitle, title, content, order, isPremium]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        bookTitle:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        title:
          type: string
        content:
          type: string
        order:
          type: integer
        isPremium:
          type: boolean
        wordCount:
          type: integer
        prevChapterId:
          type: ["integer", "null"]
        nextChapterId:
          type: ["integer", "null"]
        createdAt:
          type: string
    ChapterInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        isPremium:
          type: boolean
        order:
          type: integer
    ChapterUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPremium:
          type: boolean
        order:
          type: integer

    # ── COMMENT SCHEMAS ───────────────────────────────────
    Comment:
      type: object
      required: [id, chapterId, userId, username, content, likes, createdAt]
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        userAvatar:
          type: ["string", "null"]
        content:
          type: string
        likes:
          type: integer
        isLiked:
          type: ["boolean", "null"]
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000

    # ── USER SCHEMAS ──────────────────────────────────────
...
    CheckoutInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
    CheckoutSession:
      type: object
      required: [url]
      properties:
        url:
          type: string

    # ── SEARCH SCHEMAS ────────────────────────────────────
...